Consolidate files into one
